requirements.txt
blinker==1.8.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
img2pdf==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==2.1.5
packaging==24.1
pikepdf==9.3.0
pillow==10.4.0
qrcode==8.0
Werkzeug==3.0.4
wifi-qrcode-generator==1.0.1
wrapt==1.16.0
